part one mods

+++ code/modules/reagents/chemistry/reagents/toxins.dm
@@ -434,8 +434,15 @@
 					qdel(H.wear_mask)
 					H.update_inv_wear_mask()
 				if(H.head && !(H.head.resistance_flags & ACID_PROOF))
-					to_chat(H, "<span class='danger'>Your [H.head.name] melts away!</span>")
-					qdel(H.head)
+					if(istype(H.head, /obj/item/clothing/head/mod) && ismodcontrol(H.back))
+						var/obj/item/mod/control/C = H.back
+						var/name = H.head.name
+						C.seal_part(H.head, FALSE)
+						C.retract(null, H.head)
+						to_chat(H, "<span class='danger'>Your [name] melts away as your [C.name] performs emergency cleaning on the helmet, deactivating the suit!</span>")
+					else
+						to_chat(H, "<span class='danger'>Your [H.head.name] melts away!</span>")
+						qdel(H.head)
 					H.update_inv_head()
 				return
 		else
